Refactoring util.py into a subpackage and splitting components into a few different files. The purpose of this is so that CSVReader can be moved out of configuration.py and into the util subpackage. This will avoid some hard-to-debug circular import issues with components that need the CSVReader. One place where this is true is in statistics.py, where a plain csv.DictReader is used instead of CSVReader as a circular import workaround. These kinds of situations can snowball and affect maintainability.
Important changes to CSVReader are also in this commit:
- Exceptions related to CSV structure and formatting are now of type csv.Error
- An exception is raised if a CSV has a header but no non-whitespace rows
- An exception is raised if a row's column count differs from the others
- The function that detects all-whitespace rows has been cleaned up
